requirements.txt
anthropic>=0.18.0
openai>=1.0.0
sentence-transformers>=2.2.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
pydantic>=2.0.0
fastapi>=0.100.0
uvicorn>=0.20.0
